rename xdr2json package to xdrjson
