+++ lib/xdrjson/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "xdr2json"
+name = "xdrjson"
 version = "22.0.0"
 publish = false
 edition = "2021"
