+++ xdrjson/conversion.go
@@ -1,16 +1,16 @@
 //nolint:lll
-package xdr2json
+package xdrjson
 
 /*
 // See preflight.go for add'l explanations:
 // Note: no blank lines allowed.
 #include <stdlib.h>
-#include "../lib/xdr2json.h"
-#cgo windows,amd64 LDFLAGS: -L${SRCDIR}/libs/x86_64-pc-windows-gnu -lxdr2json -lntdll -static -lws2_32 -lbcrypt -luserenv
-#cgo darwin,amd64  LDFLAGS: -L${SRCDIR}/libs/x86_64-apple-darwin -lxdr2json -ldl -lm
-#cgo darwin,arm64  LDFLAGS: -L${SRCDIR}/libs/aarch64-apple-darwin -lxdr2json -ldl -lm
-#cgo linux,amd64   LDFLAGS: -L${SRCDIR}/libs/x86_64-unknown-linux-gnu -lxdr2json -ldl -lm
-#cgo linux,arm64   LDFLAGS: -L${SRCDIR}/libs/aarch64-unknown-linux-gnu -lxdr2json -ldl -lm
+#include "../lib/xdrjson.h"
+#cgo windows,amd64 LDFLAGS: -L${SRCDIR}/libs/x86_64-pc-windows-gnu -lxdrjson -lntdll -static -lws2_32 -lbcrypt -luserenv
+#cgo darwin,amd64  LDFLAGS: -L${SRCDIR}/libs/x86_64-apple-darwin -lxdrjson -ldl -lm
+#cgo darwin,arm64  LDFLAGS: -L${SRCDIR}/libs/aarch64-apple-darwin -lxdrjson -ldl -lm
+#cgo linux,amd64   LDFLAGS: -L${SRCDIR}/libs/x86_64-unknown-linux-gnu -lxdrjson -ldl -lm
+#cgo linux,arm64   LDFLAGS: -L${SRCDIR}/libs/aarch64-unknown-linux-gnu -lxdrjson -ldl -lm
 
 */
 import "C"
@@ -18,9 +18,10 @@ import "C"
 import (
 	"encoding"
 	"encoding/json"
-	"github.com/pkg/errors"
 	"reflect"
 	"unsafe"
+
+	"github.com/pkg/errors"
 )
 
 // ConvertBytes takes an XDR object (`xdr`) and its serialized bytes (`field`)
